==> AFM/Notebooks/00_Sorting_Spindle3d_measurements_after_montages.ipynb <==
import os
import shutil

# define root, preview folder, unsorted folder with raw files and destination folders for sorting.
root = "/Volumes/TOB_WD2/Microscopy/20230929_Centrosomes_Osmotic" + "/"
unsrt_dir = root + "Spindle3D/"

preview_dir = root + "Spindle3D_Montages/"

good_dir = root + "Spindle3D/Correct/"
bad_dir = root + "Spindle3D/Incorrect/"

good_prev = preview_dir + "Correct/"
bad_prev = preview_dir + "Incorrect/"

good = []
bad = []

good_prev_list = [f for f in os.listdir(good_prev) if not f.startswith('.')]
bad_prev_list = [f for f in os.listdir(bad_prev) if not f.startswith('.')]
#unsrt_list = [f for f in os.listdir(unsrt_dir) if not f.startswith('.')]
unsrt_list = [f.name for f in os.scandir(unsrt_dir) if f.is_dir()]

for preview_img in good_prev_list:
    preview_img_name = preview_img.rsplit("-", 1)[0]
    good.append(preview_img_name)

for preview_img in bad_prev_list:
    preview_img_name = preview_img.rsplit("-", 1)[0]
    bad.append(preview_img_name)
    

for folder in unsrt_list:
        #name = file#.split(".")[0].rsplit("_",2)[0]
        if folder in good:
            shutil.move(unsrt_dir + folder, good_dir + folder)
            print("Moved {} to Correct".format(folder))
        elif folder in bad:
            shutil.move(unsrt_dir + folder, bad_dir + folder)
            print("Moved {} to Incorrect".format(folder))
        else:
            print("{} not listed.".format(folder))
        
print("Successfully sorted Spindle3d output.")
