==> AFM/Notebooks/03_Concatenate_Spindle3D_output.ipynb <==
import pandas as pd
import os

pd.options.display.max_rows
pd.set_option('display.max_colwidth', -1)

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))



Concatenating all Spindle3D output dataframes

destination_path = "/Volumes/TOB_WD2/Data_Analysis/DataFrames/MasterDataFrame_Spindle3D.csv"

# parse Plugin Measurements
in_dir_1 = "/Volumes/TOB_WD2/Image_Analysis/Spindle3D/Out_077/Correct"

def parse_spindle3d(in_dir):
    dataframes = []
    for subdir, dirs, files in os.walk(in_dir):
        for file in files:
            filepath = subdir + os.sep + file  
            if filepath.endswith(".txt"):
                single_df = pd.read_csv(filepath, sep = " ", delimiter = "\t", encoding = 'latin1')
                single_df["Cell_ID"] = subdir.rsplit("/",1)[1]
                dataframes.append(single_df)    
            else:
                pass
    df = pd.concat(dataframes)
    return df
     
df = parse_spindle3d(in_dir_1)
not_na_df = df['Path_InputImage'].notna()
df = df[not_na_df]

# post-processing

df["Spindle_DNA_Volume_Ratio"] = df.Spindle_Volume_um3 / df.Chromatin_Volume_um3

subselection = [
    "Cell_ID", 
    "Chromatin_Dilation", 
    "Chromatin_Volume_um3", 
    "DNA_Volume_Threshold", 
    "MetaphasePlate_Length_um", 
    "MetaphasePlate_Width_um", 
    "Spindle_Angle_Degrees", 
    "Spindle_Aspect_Ratio", 
    "Spindle_DNA_Volume_Ratio",
    "Spindle_Length_um", 
    "Spindle_Volume_um3", 
    "Spindle_Width_Avg_um", 
    "Tubulin_Spindle_Average_Intensity", 
    "Version"
]

df = df[subselection]
df.reset_index(inplace = True)

df.to_csv(destination_path)
print("Finished compilation of main Spindle3D dataframe: {}".format(destination_path))
