==> AFM/Notebooks/06_Time_coordinates.ipynb <==
import time
import datetime
import pandas as pd
import os

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

# Get LowZoom Timestamps via experiment Log

rootDirectory = "/Volumes/TOB_WD2/Data_Analysis/DataFrames"
inputDirectory = rootDirectory + "/Times/Experiment_logs"


def get_unixtime(x):
    unix_time = time.mktime(datetime.datetime.strptime(x, "%Y/%m/%d %H:%M:%S").timetuple())
    return unix_time

def numbering_for_conc(x):
    lowzoom_id, Timeframe_old = x.rsplit("-",1)
    Timeframe_old_int = int(Timeframe_old[1:])
    Timeframe_new = str(Timeframe_old_int + 89).zfill(4)
    fixed_id = lowzoom_id + "-T" + Timeframe_new
    return fixed_id


def read_experiment_logs(folder):
    logs = []
    for root, dirs, files in os.walk(folder):
        files = [f for f in files if not f[0] == '.']
        for log_file in files:
            print("Starting analysis of {}.".format(log_file))
            log_file_path = root + os.sep + log_file
            date_string = log_file.split("_")[0]
            date_int = int(date_string.split("-")[0])
            
            exlog = pd.read_csv(log_file_path, sep = '\t', lineterminator = '\r')
            
            exlog = exlog[~exlog['Message'].str.contains("LowZoomStitch", na = False)]
            exlog = exlog[exlog['Message'].str.contains("LowZoom", na = False)]

            if date_int < 20200806:
                exlog = exlog[exlog['Message'].str.contains("Recorded and saved image", na = False)]
            else:
                exlog = exlog[exlog['Message'].str.contains("Saved image", na = False)]
                
            exlog["LowZoom_ID"] = exlog['Message'].str.split("\\").str.get(-1).str.split(".").str.get(0)
            
            if date_string == "20190618-193725":
                exlog["LowZoom_ID"] = exlog.LowZoom_ID.apply(numbering_for_conc)
                
            if date_int == 20190618:
                exlog["LowZoom_ID"] = exlog.LowZoom_ID.apply(lambda x: "20190618-conc" + "_" + x)
        
            else: 
                exlog["LowZoom_ID"] = exlog.LowZoom_ID.apply(lambda x: date_string + "_" + x)
            exlog["Timestamp"] = exlog.Time.apply(get_unixtime)
            
            if date_int == 20190618:
                start_time = 1560879470
            elif date_int == 20200725:
                start_time = 1595614672
            elif (date_int == 20200802) or (date_int == 20200803):
                start_time = 1596211227
            elif (date_int == 20200809) or (date_int == 20200810):
                start_time = 1596814952
            
            else:
                start_time = exlog.iloc[0,-1]
            
            exlog["Experiment_StartTime_mins"] = start_time
            print(date_string + " start time: " + str(start_time))
            exlog["Experiment_Time_mins"] = (exlog.Timestamp - start_time) / 60 # Timestamps are seconds
            exlog = exlog.drop(["Message", "Type"], axis = 1)
            exlog = exlog.reset_index(drop = True)
            logs.append(exlog)
            print("Finishing analysis of {}.".format(log_file))
    
    log_df = pd.concat(logs)
    return log_df

df = read_experiment_logs(inputDirectory)

df.to_csv(rootDirectory + "/" + "MasterDataFrame_Times.csv")
print("Saved Time Dataframe in " + rootDirectory + "/" + "MasterDataFrame_Times.csv")
