==> AFM/Notebooks/10_Statistics_Population.ipynb <==
import os 
import pandas as pd
import numpy as np

from sklearn import preprocessing
from sklearn import linear_model
from sklearn.metrics import mean_squared_error, r2_score
import sklearn.linear_model

import scipy
from scipy import stats
from scipy.optimize import curve_fit

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

root_dir = "/Volumes/TOB_WD2/Data_Analysis/DataFrames" + "/"

# load data

df = pd.read_csv(root_dir + "MasterDataFrame_Population_Morphometry_FilteredForStatistics.csv") # only Spindle3D overlap 
#df = pd.read_csv(root_dir + "MasterDataFrame_Population_Morphometry.csv") # not filtered


df_p1 = df[df.Condition == "1_Pluripotent"]
df_p2 = df[df.Condition == "2_Differentiation"]

df.groupby("Condition").Differentiation_bins.value_counts() ##### always check whether filtered or not! #####

df.groupby(["Cell_Volume_bin", "Condition"]).Cell_Volume_um3.mean()

df.groupby(["Cell_Volume_bin", "Condition"]).Cell_Volume_um3.std()

df.groupby(["Cell_Volume_bin", "Condition"]).Spindle_Volume_um3.mean()

df.groupby(["Cell_Volume_bin", "Condition"]).Spindle_Volume_um3.std()

df.groupby(["Cell_Volume_bin", "Condition"]).Fraction_Tubulin_in_Spindle.mean()

df.groupby(["Cell_Volume_bin", "Condition"]).Fraction_Tubulin_in_Spindle.std()


df.groupby(["Differentiation_bins", "Condition"]).Cell_Volume_um3.mean()

df.groupby(["Differentiation_bins", "Condition"]).Cell_Volume_um3.std()

df.groupby(["Differentiation_bins", "Condition"]).SurfaceArea.mean()

df.groupby(["Differentiation_bins", "Condition"]).SurfaceArea.std()

df.groupby(["Differentiation_bins", "Condition"]).Spindle_Volume_um3.mean()

df.groupby(["Differentiation_bins", "Condition"]).Spindle_Volume_um3.std()

df.groupby(["Differentiation_bins", "Condition"]).Fraction_Tubulin_in_Spindle.mean()

df.groupby(["Differentiation_bins", "Condition"]).Fraction_Tubulin_in_Spindle.std()

df.groupby(["Differentiation_bins", "Condition"]).Fraction_SpindleVol_in_Cell.mean()


df.groupby(["Differentiation_bins", "Condition"]).Fraction_SpindleVol_in_Cell.std()

df.groupby(["Differentiation_bins", "Condition"]).Tubulin_density_spindle_norm.mean()

df.groupby(["Differentiation_bins", "Condition"]).Tubulin_density_spindle_norm.std()

df.groupby(["Differentiation_bins", "Condition"]).Tubulin_density_cytop_norm.mean()

df.groupby(["Differentiation_bins", "Condition"]).Tubulin_density_cytop_norm.std()

# Linear Regression

# also test regression via scipy.stats.linregress(x, y) !!!

def linear_fit(dataframe_list, independent_column, dependent_column): 
    for dataframe in dataframe_list:
        dataframe = dataframe[~dataframe[independent_column].isna() & ~dataframe[dependent_column].isna()]
        
        dataframe_name = dataframe.Condition.head(1)
        length = dataframe.shape[0]

        X = dataframe[independent_column].values.reshape(length, 1)
        y = dataframe[dependent_column].values.reshape(length, 1)

        regr = linear_model.LinearRegression()
        regr.fit(X, y)
        y_predicted = regr.predict(X)

        # model evaluation
        rmse = mean_squared_error(y, y_predicted)
        R2 = r2_score(y, y_predicted)
        slope = regr.coef_
        interc = regr.intercept_

        print(
            "The linear fit for x = {} vs y = {} in {}: y = {} x +- {}. R2 = {}, RMSE = {}.".format(
                dependent_column, 
                independent_column,
                dataframe_name, slope, interc, R2, rmse
            )
        )
    return rmse, R2, slope, interc

# Non - linear regression

def power_law(x, a, b):
    # x = a * x ^ b
    return a * np.power(x, b)

# return parameters (pars) a and b, and the covariance (cov)
def curve_fitting(dataframe, independent_variable, dependent_variable):
    dataframe = dataframe[[independent_variable, dependent_variable]].dropna()
    pars, cov = curve_fit(
        f = power_law, 
        ydata = dataframe[dependent_variable], 
        xdata = dataframe[independent_variable], 
        p0 = [0, 0], 
        bounds = (-np.inf, np.inf), 
        absolute_sigma = True
    )
    # Get the standard deviations of the parameters (square roots of the # diagonal of the covariance)
    # SEEMS LIKE THIS IS THE STANDARD ERROR? 
    stdevs = np.sqrt(np.diag(cov))
    # Calculate the residuals
    res = dataframe[dependent_variable] - power_law(dataframe[independent_variable], *pars)
    print("Power law between {} (independent var) and {}, factor a = ".format(independent_variable, dependent_variable) + str(pars[0]) + " +- " + str(stdevs[0]))
    print("exponent b= " + str(pars[1]) + " +- " + str(stdevs[1]))
    return pars, cov, stdevs, res

# correlation
def correlation(dataframe_list, independent_column, dependent_column):
    for dataframe in dataframe_list:
        dataframe_name = dataframe.Condition.head(1)
        spearman = stats.spearmanr(
            dataframe[independent_column], 
            dataframe[dependent_column], 
            nan_policy = 'omit')
        print(
            "Spearman correlation for x = {} vs y = {} in {}: {}.".format(
                dependent_column, 
                independent_column,
                dataframe_name, spearman
            )
        )
        
def pearson_correlation(dataframe_list, independent_column, dependent_column):
    for dataframe in dataframe_list:
        dataframe_name = dataframe.Condition.head(1)
        spearman = stats.pearsonr(
            dataframe[independent_column], 
            dataframe[dependent_column], 
            )
        print(
            "Pearson correlation for x = {} vs y = {} in {}: {}.".format(
                dependent_column, 
                independent_column,
                dataframe_name, spearman
            )
        )

# bootstrapping correlations

def bootstrap_correlation(dataframe_list, variable_1, variable_2, iterations = 1000, n_subsampling = 500):

    corr_dataframe_list = []
    for dataframe in dataframe_list:
        dataframe = dataframe[[variable_1, variable_2, "Condition"]].dropna()
        condition = dataframe.Condition.head(1).item()
        print(condition)
        coefficient_lst = []
        for i in range(iterations):
            data_sample = dataframe.sample(n_subsampling, replace = True)
            spearman = data_sample[variable_1].corr(
                data_sample[variable_2], 
                method = 'spearman'
            )
            coefficient_lst.append(spearman)
            print(
                "Sampled Spearman's correlation coefficient: {}".format(spearman)
            )
        coefficient_df = pd.DataFrame(coefficient_lst)
        coefficient_df["Condition"] = condition
        corr_dataframe_list.append(coefficient_df)
    df = pd.concat(corr_dataframe_list)
    df = df.rename({0: 'Spearman_coefficient'.format(variable_1, variable_2)}, axis = 1) 
    df["Colour"] = df.Condition.apply(
        lambda x: "#274754" if x == "2_Differentiation" else "#E8C468"
    )
    print("Concatenated dataframe containing bootstrapped correlation coefficients.")
    df.to_csv(root_dir + 'BOOTSTRAP_Spearman_{}_{}.csv'.format(variable_1, variable_2))
    print("Saved dataframe in: " + root_dir)
    return df

SVdensity_bootstrap = bootstrap_correlation(
    dataframe_list = [df_p1, df_p2], 
    variable_1 = "Tubulin_density_spindle_norm", 
    variable_2 = "Spindle_Volume_um3", 
    iterations = 1000, 
    n_subsampling = 500
)

# coefficient of variation

def CoV(measurement, condition_1, condition_2):
    CoV_p1 = scipy.stats.variation(condition_1[measurement], nan_policy = 'omit')
    CoV_p2 = scipy.stats.variation(condition_2[measurement], nan_policy = 'omit')
    print("The coefficient of variation for {} in PLURIPOTENT cells is: ".format(measurement) + str(CoV_p1))
    print("The coefficient of variation for {} in DIFF cells is: ".format(measurement) + str(CoV_p2))


# t-test

def ttest(measurement):
    statistic, pvalue = scipy.stats.ttest_ind(
        df_p1[measurement], 
        df_p2[measurement], 
        axis = 0, 
        equal_var = False, 
        nan_policy = 'omit'
    )
    print ("The p-value for {} is: ".format(measurement) + str(pvalue))

def ttest_binned(dependent_measurement, binned_measurement):
    bin_list = df[binned_measurement].unique()
    for binning in bin_list:
        df_p1_binning = df_p1[df_p1[binned_measurement] == binning]
        df_p2_binning = df_p2[df_p2[binned_measurement] == binning]   
        
        statistic, pvalue = scipy.stats.ttest_ind(
            df_p1_binning[dependent_measurement], 
            df_p2_binning[dependent_measurement], 
            axis = 0, 
            equal_var = False, 
            nan_policy = 'omit'
        )
        print ("The p-value for {} in the bin {} is: ".format(dependent_measurement, binning) + str(pvalue))
        

ttest_binned(
    "Chromatin_Volume_um3", 
    "Differentiation_bins" # binned measurement
)

df.groupby(["Differentiation_bins"]).Condition.value_counts()

ttest_binned(
    "Fraction_Tubulin_in_Spindle", 
    "Cell_Volume_bin" # binned measurement
)

df.groupby(["Cell_Volume_bin"]).Condition.value_counts()

ttest("Chromatin_Volume_um3")

Spearman_SV_TubDens = correlation(
    dataframe_list = [df_p1, df_p2], 
    dependent_column = "Spindle_Volume_um3", 
    independent_column = "Tubulin_density_spindle_norm"
)

Spearman_SM_TubDens = correlation(
    dataframe_list = [df_p1, df_p2], 
    dependent_column = "Tubulin_mass_spindle_norm", 
    independent_column = "Tubulin_density_spindle_norm"
)
Spearman_SM_TubDens

Lfit_SV_CV = linear_fit(
    dataframe_list = [df_p1, df_p2], 
    dependent_column = "Spindle_Volume_um3", 
    independent_column = "Cell_Volume_um3"
)

Lfit_SM_CV = linear_fit(
    dataframe_list = [df_p1, df_p2], 
    dependent_column = "Tubulin_mass_spindle_norm", 
    independent_column = "Cell_Volume_um3"
)

POWER_SACV_SV_p1 = curve_fitting(
    df[df.Condition == "1_Pluripotent"], 
    "CellSurfaceArea_CellVolume_Ratio", 
    "Spindle_Volume_um3"
)
POWER_SACV_SV_p2 = curve_fitting(
    df[df.Condition == "2_Differentiation"], 
    "CellSurfaceArea_CellVolume_Ratio", 
    "Spindle_Volume_um3"
)
print(POWER_SACV_SV_p1, POWER_SACV_SV_p2)  

POWER_SACV_SM_p1 = curve_fitting(
    df[df.Condition == "1_Pluripotent"], 
    "CellSurfaceArea_CellVolume_Ratio", 
    "Tubulin_mass_spindle_norm"
)
POWER_SACV_SM_p2 = curve_fitting(
    df[df.Condition == "2_Differentiation"], 
    "CellSurfaceArea_CellVolume_Ratio", 
    "Tubulin_mass_spindle_norm"
)
print(POWER_SACV_SM_p1, POWER_SACV_SM_p2)

POWER_TubDensity_SSR_p1 = curve_fitting(
    df[df.Condition == "1_Pluripotent"], 
    "Tubulin_density_spindle_norm", 
    "Fraction_SpindleVol_in_Cell"
)
POWER_TubDensity_SSR_p2 = curve_fitting(
    df[df.Condition == "2_Differentiation"], 
    "Tubulin_density_spindle_norm", 
    "Fraction_SpindleVol_in_Cell"
)
print(POWER_TubDensity_SSR_p1, POWER_TubDensity_SSR_p2)

ttest("Tubulin_density_spindle_norm")

# Testing for significantly different variances
# without the need for normal distribution

from scipy.stats import bartlett
from scipy.stats import levene

stat_bartlett, p_bartlett = bartlett(df_p1.Tubulin_density_spindle_norm, df_p2.Tubulin_density_spindle_norm)
print(p_bartlett)

stat_levene, p_levene = levene(df_p1.Tubulin_density_spindle_norm, df_p2.Tubulin_density_spindle_norm)
print(stat_levene)

print(df_p1.Tubulin_density_spindle_norm.var())
print(df_p2.Tubulin_density_spindle_norm.var())
print(df_p1.Tubulin_density_spindle_norm.std())
print(df_p2.Tubulin_density_spindle_norm.std())
print(df_p1.Tubulin_density_spindle_norm.mean())
print(df_p2.Tubulin_density_spindle_norm.mean())

df.groupby("Condition").Cell_Volume_um3.mean()

df.groupby(["Condition", "Cell_Volume_bin"]).Cell_Volume_um3.mean()

df.groupby("Condition").Spindle_Volume_um3.mean()

df.groupby("Condition").Fraction_SpindleVol_in_Cell.mean()

df.groupby("Condition").Fraction_SpindleVol_in_Cell.std()

df.groupby("Condition").Chromatin_Volume_um3.mean()

df.groupby("Condition").Chromatin_Volume_um3.std()

df.groupby(["Condition", "Cell_Volume_bin"]).Spindle_Volume_um3.mean()

df.groupby(["Condition", "Cell_Volume_bin"]).Spindle_Volume_um3.median()

df.groupby(["Condition", "Cell_Volume_bin"]).Spindle_Length_um.median()

df.groupby(["Condition", "Cell_Volume_bin"]).Tubulin_mass_spindle_norm.mean()

df.groupby("Condition").Fraction_SpindleVol_in_Cell.median()

df.groupby("Condition").Fraction_Tubulin_in_Spindle.mean()

df.groupby("Condition").Fraction_Tubulin_in_Spindle.std()

df.groupby("Condition").Tubulin_density_spindle_norm.mean()

df.groupby("Condition").Tubulin_density_spindle_norm.std()

df.groupby("Condition").Tubulin_density_cytop_norm.mean()

correlation([df_p1, df_p2], "Cell_Volume_um3", "Spindle_Volume_um3")

correlation([df], "Cell_Volume_um3", "Spindle_Volume_um3")

correlation([df], "Tubulin_mass_spindle_norm", "Chromatin_Volume_um3")

correlation([df], "Tubulin_mass_spindle_norm", "Tubulin_density_spindle_norm")

correlation([df], "Tubulin_mass_spindle_norm", "Spindle_Volume_um3")

correlation([df], "Tubulin_density_cytop_norm", "Cell_Volume_um3")

#pearson_correlation([df], "Cell_Volume_um3", "Spindle_Volume_um3")

from numpy import var, mean, sqrt
from pandas import Series

def cohens_d(parameter) -> float:
    d1 = df_p1[parameter]
    d2 = df_p2[parameter]
    
    # calculate the size of samples
    n1, n2 = len(d1), len(d2)

    # calculate the variance of the samples
    s1, s2 = var(d1, ddof = 1), var(d2, ddof = 1)

    # calculate the pooled standard deviation
    s = sqrt(((n1 - 1) * s1 + (n2 - 1) * s2) / (n1 + n2 - 2))

    # calculate the means of the samples
    u1, u2 = mean(d1), mean(d2)

    # return the effect size
    return (u1 - u2) / s

def cohens_d_binned(parameter, binned_parameter):
    
    bin_list = df[binned_parameter].unique()
    for binning in bin_list:
        print(binning)
        d1 = df_p1[df_p1[binned_parameter] == binning][parameter]
        d2 = df_p2[df_p2[binned_parameter] == binning][parameter]
        # calculate the size of samples
        n1, n2 = len(d1), len(d2)
        # calculate the variance of the samples
        s1, s2 = var(d1, ddof = 1), var(d2, ddof = 1)
        # calculate the pooled standard deviation
        s = sqrt(((n1 - 1) * s1 + (n2 - 1) * s2) / (n1 + n2 - 2))
        # calculate the means of the samples
        u1, u2 = mean(d1), mean(d2)
        # return the effect size
        cohens_d = ((u1 - u2) / s)
        print(cohens_d)
    #return cohens_d

cohens_d("Fraction_SpindleVol_in_Cell")

cohens_d("Fraction_Tubulin_in_Spindle")

cohens_d("Sphericity")

cohens_d("Tubulin_density_cytop_norm")

cohens_d("Tubulin_density_spindle_norm")

cohens_d("Chromatin_Volume_um3")

cohens_d_binned("Fraction_SpindleVol_in_Cell", "Cell_Volume_bin")

cohens_d_binned("Fraction_SpindleVol_in_Cell", "Differentiation_bins")

cohens_d_binned("Sphericity", "Differentiation_bins")

cohens_d_binned("Cell_Volume_um3", "Differentiation_bins")

cohens_d_binned("Chromatin_Dilation", "Differentiation_bins")

cohens_d_binned("MetaphasePlate_Width_um", "Differentiation_bins")

cohens_d_binned("Spindle_Volume_um3", "Cell_Volume_bin")

cohens_d_binned("Spindle_Volume_um3", "CellSurfaceArea_CellVolume_bin")

df_select = df[["Cell_Volume_um3", "Differentiation_bins", "Condition"]]
df_select["Differentiation_bins_Condition"] = df_select.Differentiation_bins.str.cat(df_select.Condition)
print(df_select.Differentiation_bins_Condition.value_counts())
df_select.sample(30)

df_select.sample(30)

# ANOVA Testing for time-resolved / binned data!
# ANOVA as generalized linear model (GLM):
import numpy as np
import statsmodels.api as sm
from statsmodels.formula.api import ols
import pingouin as pg
import scipy.stats as stats
from statsmodels.stats.multicomp import pairwise_tukeyhsd
from statsmodels.stats.multicomp import MultiComparison

measurement = 'Sphericity'
group_variable_string = 'Differentiation_bins'
group_variable = group_variable_string + '_Condition'

df[group_variable] = df[group_variable_string].str.cat(df.Condition)
 
nan_elements = df[measurement].isnull()
data = df[~nan_elements]

list_of_groups = data[group_variable].unique()

print(data[group_variable].value_counts())

# Bartlett's test for equal variances (One-way ANOVA requires equal variances!)

BartlettResult = stats.bartlett(
    data[data[group_variable] == list_of_groups[0]][measurement], 
    data[data[group_variable] == list_of_groups[1]][measurement], 
    data[data[group_variable] == list_of_groups[2]][measurement],
    data[data[group_variable] == list_of_groups[3]][measurement], 
    data[data[group_variable] == list_of_groups[4]][measurement], 
    data[data[group_variable] == list_of_groups[5]][measurement],
)

print("The Bartlett test for equal variances of {}: ".format(measurement) + str(BartlettResult))

results = ols(measurement + '~ C('+ group_variable + ')', data = data).fit()
print(results.summary())

aov_table = sm.stats.anova_lm(results, typ = 2)

def anova_table(aov):
    aov['mean_sq'] = aov[:]['sum_sq'] / aov[:]['df']
    
    aov['eta_sq'] = aov[:-1]['sum_sq'] / sum(aov['sum_sq'])
    
    aov['omega_sq'] = (aov[:-1]['sum_sq']-(aov[:-1]['df'] * aov['mean_sq'][-1])) / (sum(aov['sum_sq']) + aov['mean_sq'][-1])
    
    cols = ['sum_sq', 'df', 'mean_sq', 'F', 'PR(>F)', 'eta_sq', 'omega_sq']
    aov = aov[cols]
    return aov

aov_table = anova_table(aov_table)
print("\n ANOVA TABLE: ")
print(aov_table)

# Post-hoc testing
mc = MultiComparison(data[measurement], data[group_variable])
mc_results = mc.tukeyhsd()
print("\n\n POST-HOC testing for {}: \n".format(measurement))
print(mc_results)
print("If \"reject\" = True, then H0 should be rejected")

# Welch's ANOVA when variances are unequal
aov_table_WELCH = pg.welch_anova(dv = measurement, between = group_variable, data = data)
print("\n Welch's ANOVA table") 
print(aov_table_WELCH)

# Post-hoc testing using Games-Howell post-hoc test
mc_results_GamesHowell = pg.pairwise_gameshowell(dv = measurement, between = group_variable, data = data)
print("\n Games-Howell post-hoc test table") 
print(mc_results_GamesHowell)
