==> AFM/Notebooks/12_Coordinate_Matching.ipynb <==
import pandas as pd
pd.set_option("display.max_columns", None)
import numpy as np
import os

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

root = "/Volumes/TOB_WD2/Data_Analysis/DataFrames" + "/"
lineageDir = root + "Tracking/Output_Lineages" + "/"

lineages = []

for subdir, dirs, files in os.walk(lineageDir):
            for file in files:
                if not file.startswith('.'):
                    filepath = subdir + os.sep + file  
                    if filepath.endswith(".csv"):
                        single_df = pd.read_csv(filepath)
                        lineages.append(single_df)    

lineage_df = pd.concat(lineages)  
lineage_df = lineage_df.loc[:, ~lineage_df.columns.str.contains('^Unnamed')]

df = pd.read_csv(root + "MasterDataFrame_Population_Morphometry.csv")

df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
lineage_df.tail(5)

def convert_TMcoordinates(x, coord):
    # convert uncalibrated trackmate coordinates to calibrated ones
    if not (x.Dataset == 20190618) & (x.Position == 2):
        if coord == "X":
            return x.Track_Coordinate_X
        if coord == "Y":
            return x.Track_Coordinate_Y
    else:
        px_per_micron = 1.61
        if coord == "X":
            Track_Coordinate_X_calibrated = x.Track_Coordinate_X / px_per_micron
            return Track_Coordinate_X_calibrated
        else:
            Track_Coordinate_Y_calibrated = x.Track_Coordinate_Y / px_per_micron
            return Track_Coordinate_Y_calibrated

lineage_df["Track_Coordinate_X"] = lineage_df.apply(convert_TMcoordinates, coord = "X", axis = 1)
lineage_df["Track_Coordinate_Y"] = lineage_df.apply(convert_TMcoordinates, coord = "Y", axis = 1)

lineage_df.head()

def convert_timepoints(x):
    # adjust timepoints for interrupted experiments CHECK 0618!!!
    if x.Experiment == "20200725-142832":
        return x.Timepoint + 93
    elif x.Experiment == "20200802-104739":
        return x.Timepoint + 225
    elif x.Experiment == "20200803-181830":
        return x.Timepoint + 225 + 149
    elif x.Experiment == "20200809-114236":
        return x.Timepoint + 199
    elif x.Experiment == "20200810-225504":
        return x.Timepoint + 199 + 179
    else:
        return x.Timepoint

def convert_coordinates(x, coord):
    # convert pixel coordinates to um
    if x.Experiment == "20190618-conc":
        px_per_micron = 1.61
        if coord == "X":
            X_um = x.X_px / px_per_micron
            return X_um
        else:
            Y_um = x.Y_px / px_per_micron
            return Y_um
    else:
        px_per_micron = 3.22
        if coord == "X":
            X_um = x.X_px / px_per_micron
            return X_um
        else:
            Y_um = x.Y_px / px_per_micron
            return Y_um
        


        
df["X_um"] = df.apply(convert_coordinates, coord = "X", axis = 1)
df["Y_um"] = df.apply(convert_coordinates, coord = "Y", axis = 1)
df["Timepoint"] = df.apply(convert_timepoints, axis = 1)

df.head()

def find_match(x, trackmate, dataframes_list):
    # matching based on µm distances!
        cell_id = x.Cell_ID
        match = trackmate.loc[(trackmate.Position == x.Position) & (trackmate.Dataset == x.Dataset) &
                                (np.isclose(np.array(trackmate.Timepoint), np.array(x.Timepoint),\
                                              atol = 4.0) == np.array([ True ])) & 
                                (np.isclose(np.array(trackmate.Track_Coordinate_X), np.array(x.X_um),\
                                              atol = 10.0 ) == np.array([ True ])) & 
                                (np.isclose(np.array(trackmate.Track_Coordinate_Y), np.array(x.Y_um),\
                                              atol = 10.0 ) == np.array([ True ]))
                                ,:]
        if match.shape[0] > 0:
                match['Delta_x'] = match.Track_Coordinate_X.apply(lambda coord_x: abs(coord_x - x.X_um))
                match['Delta_y'] = match.Track_Coordinate_Y.apply(lambda coord_y: abs(coord_y - x.Y_um))
                match['Delta_t'] = match.Timepoint.apply(lambda time: abs(time - x.Timepoint))
                match.sort_values(by = ['Delta_t', 'Delta_x', 'Delta_y'], inplace = True)
                match = match.head(1)
                match["Cell_ID"] = cell_id
                dataframes_list.append(match)
        else:
            pass

matches_list = []
df.apply(find_match, trackmate = lineage_df, dataframes_list = matches_list, axis = 1)
matches = pd.concat(matches_list, axis = 0, ignore_index = True)
matches.shape

matches.head()

def matching_statistics(data, lineage_data, matches_data):
    list_of_datasets = lineage_data.Dataset.unique()
    data_count = data.groupby(["Dataset", "Position"]).Cell_ID.count().reset_index()
    lineage_count = lineage_data.groupby(["Dataset", "Position"]).Source_ID.count().reset_index()
    matches_count = matches_data.groupby(["Dataset", "Position"]).Cell_ID.count().reset_index()
    
    delta_x_avg = matches_data.groupby(["Dataset", "Position"]).Delta_x.mean().reset_index()
    delta_y_avg = matches_data.groupby(["Dataset", "Position"]).Delta_y.mean().reset_index()
    delta_t_avg = matches_data.groupby(["Dataset", "Position"]).Delta_t.mean().reset_index()

    dictionaries = []
    for dataset in list_of_datasets:
        list_of_positions = lineage_data[lineage_data.Dataset == dataset].Position.unique()
        for position in list_of_positions:
            no_of_cell_ids = data_count.loc[
                (data_count.Dataset == int(dataset)) & (data_count.Position == int(position)), "Cell_ID"
            ]
            if no_of_cell_ids.shape[0] > 0:
                no_of_cell_ids = no_of_cell_ids.item()
            else:
                no_of_cell_ids = 0
            
            no_of_splitting_events = lineage_count.loc[
                (lineage_count.Dataset == int(dataset)) & (lineage_count.Position == int(position)), "Source_ID"
            ]
            if  no_of_splitting_events.shape[0] > 0:
                no_of_splitting_events = no_of_splitting_events.item()
            else:
                no_of_splitting_events = 0
            
            no_of_matches = matches_count.loc[
                (matches_count.Dataset == int(dataset)) & (matches_count.Position == int(position)), "Cell_ID"
            ]
            if no_of_matches.shape[0] > 0:
                no_of_matches = no_of_matches.item()
            else:
                no_of_matches = 0
            
            average_delta_x = delta_x_avg.loc[
                (matches_count.Dataset == int(dataset)) & (matches_count.Position == int(position)), "Delta_x"
            ]
            if average_delta_x.shape[0] > 0:
                average_delta_x = average_delta_x.item()
            else:
                average_delta_x = None
                
            average_delta_y = delta_y_avg.loc[
                (matches_count.Dataset == int(dataset)) & (matches_count.Position == int(position)), "Delta_y"
            ]
            if average_delta_y.shape[0] > 0:
                average_delta_y = average_delta_y.item()
            else:
                average_delta_y = None
                
            average_delta_t = delta_t_avg.loc[
                (matches_count.Dataset == int(dataset)) & (matches_count.Position == int(position)), "Delta_t"
            ]
            if average_delta_t.shape[0] > 0:
                average_delta_t = average_delta_t.item()
            else:
                average_delta_t = None

            sub_dictionary = {"Dataset": dataset, 
                              "Position": position, 
                              "Number of Cell IDs": no_of_cell_ids, 
                              "Number of splitting events": no_of_splitting_events, 
                              "Number of matches": no_of_matches, 
                              "Maximum yield": no_of_cell_ids / no_of_splitting_events, 
                              "Matches per split event": no_of_matches / no_of_splitting_events, 
                              "Matches per cell id": no_of_matches / no_of_cell_ids, 
                              "Avg Delta x": average_delta_x, 
                              "Avg Delta y": average_delta_y, 
                              "Avg Delta t": average_delta_t}
            sub_df = pd.DataFrame(sub_dictionary, index = [0])
            dictionaries.append(sub_df)
    return pd.concat(dictionaries)
    
matching_stats = matching_statistics(data = df, lineage_data = lineage_df, matches_data = matches)

matching_stats.to_csv("/Volumes/TOB_WD2/Data_Analysis/DataFrames/MetaStatistics/MetaStatistics_CoordinateMatching.csv")
print("Finished saving the metastatistics for tracking.")

matching_stats

# Merge and export
final_df = matches.merge(df, on = "Cell_ID", how = "inner")
final_df.drop(["Position_x", "Experiment_x", "Timepoint_x", "Dataset_x", "index"], axis = 1, inplace = True)
final_df.rename(columns = {
    "Position_y": "Position", 
    "Experiment_y": "Experiment", 
    "Timepoint_y": "Timepoint", 
    "Dataset_y": "Dataset"
}, inplace = True)

final_df.to_csv(root + "MasterDataFrame_MatchesLineages.csv")
print("Finished exporting the final dataframe.")
