==> AFM/Notebooks/13_CellCycles.ipynb <==
import pandas as pd
import os
import math

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

root = "/Volumes/TOB_WD2/Data_Analysis/DataFrames" + "/"
lineages = root + "/Tracking/Output_Lineages" + "/"

def conc_lineage_df(in_dir):
    dataframes = []
    for subdir, dirs, files in os.walk(in_dir):
        for file in files:
            if not file.startswith("."):
                filepath = subdir + os.sep + file  
                if filepath.endswith(".csv"):
                    single_df = pd.read_csv(filepath)
                    dataframes.append(single_df)    
                else:
                    pass
            else:
                pass
    df = pd.concat(dataframes)
    return df
     
lineages_df = conc_lineage_df(lineages)
times_df = pd.read_csv(root + "MasterDataFrame_Times.csv")

# Preparing times for merging 
def parent_dataset(x):
        datasets = [
            "20190227-182447", 
            "20190618-conc", 
            "20190827-181922",
            "20200724-201611",
            "20200728-174144", 
            "20200730-190931",
            "20200731-175845",
            "20200806-164803",
            "20200807-174159"
        ]

        if x in datasets:
            return x.split("-")[0]

        elif (x == "20200802-104739") or (x == "20200803-181830"):
            return "20200731"
        elif x == "20200725-142832":
            return "20200724"
        elif (x == "20200809-114236") or (x == "20200810-225504"):
            return "20200807"
        else:
            print("Does not apply?: " + x)

def correct_timepoints(x):
    subexperiment_key = x.Subexperiment
    timepoint = x.Timepoint
    timepoint_dir = {"20200725-142832": 93, 
                     "20200802-104739": 225, 
                     "20200803-181830": 374, 
                     "20200809-114236": 199, 
                     "20200810-225504": 378
                    }
    if subexperiment_key in timepoint_dir:
        new_timepoint = timepoint + timepoint_dir[subexperiment_key]
        return new_timepoint
    else:
        return timepoint

            
lineages_df.rename(columns = {"Experiment": "Subexperiment"}, inplace = True)
lineages_df["Experiment"] = lineages_df.Subexperiment.apply(parent_dataset)
print(lineages_df.Experiment.unique())
times_df["Subexperiment"] = times_df.LowZoom_ID.str.split("_").str.get(0)
times_df["Experiment"] = times_df.Subexperiment.apply(parent_dataset)  

times_df["Timepoint"] = times_df.LowZoom_ID.apply(lambda x: int(x[-4:]))
times_df["Timepoint"] = times_df.apply(correct_timepoints, axis = 1)
times_df["Position"] = times_df.LowZoom_ID.apply(lambda x: int(x[-7]))
destination_times = root + "MasterDataFrame_Times_II.csv"
times_df.to_csv(destination_times)

print(lineages_df.Subexperiment.unique())
print(times_df.Subexperiment.unique())

df = times_df.merge(lineages_df, how = "inner", on = ["Timepoint", "Experiment", "Subexperiment", "Position"]) 
df = df.loc[:, ~df.columns.str.contains('^Unnamed')]

destination = root + "MasterDataFrame_MergeLineagesTimes.csv"
df.to_csv(destination)

# Quality control meta statistics after merging 
# Times and Lineages

from statistics import mean

def get_lineage_statistics(dataframe = df):
    dataframelist = dataframe.Experiment.unique()
    positions = [1, 2]
    dataframes_for_concat = []
    for data_name in dataframelist:
        for position in positions:    

            data = dataframe[(dataframe["Experiment"] == data_name) & (dataframe["Position"] == position)]
            dataset = data_name
            position = position
            max_generation = data.Generation.max()
            No_of_source_ids = data.Source_ID.dropna().shape[0]
            No_of_sister_ids = data.Sister_ID.dropna().shape[0]
            No_of_mother_ids = data.Mother_ID.dropna().shape[0]
            No_of_aunt_ids = data.Aunt_ID.dropna().shape[0]
            No_of_cousin_ids = data.Cousin_ID.dropna().shape[0]
            No_of_grandmother_ids = data.Grandmother_ID.dropna().shape[0]

            statistics_dict = {"Dataset": dataset, 
                               "Position": position, 
                               "Max_No_of_Generations": max_generation,
                               "Number_of_Source_IDs": No_of_source_ids,
                               "Number_of_Sister_IDs": No_of_sister_ids,
                               "Number_of_Cousin_IDs": No_of_cousin_ids,
                               "Number_of_Mother_IDs": No_of_mother_ids,
                               "Number_of_Aunt_IDs": No_of_aunt_ids,
                               "Number_of_Grandmother_IDs": No_of_grandmother_ids,
                              }
            df = pd.DataFrame(statistics_dict, index = [0])
            dataframes_for_concat.append(df)
    
    final_df = pd.concat(dataframes_for_concat)
    return final_df

statistics_df = get_lineage_statistics()
statistics_df

def get_cell_cycle(x, dataframe):
    daughter_time = x.Experiment_Time_mins
    if x.Generation < 2:
        return None
    else:
        mother_row = dataframe.loc[
            (dataframe.Experiment == x.Experiment) &
 #           (dataframe.Subexperiment == x.Subexperiment) & #mothers can be in older subexperiments!
            (dataframe.Position == x.Position) &
            (dataframe.Source_ID == x.Mother_ID)
        ]
        if mother_row["Experiment_Time_mins"].shape[0] == 1:   
            mother_time = mother_row["Experiment_Time_mins"].item()
            cell_cycle = daughter_time - mother_time
            return cell_cycle
        else:
            return None

df["Cell_Cycle_mins"] = df.apply(get_cell_cycle, dataframe = df, axis = 1)

print("Finished calculating cell cycles.")
df.tail()

df.columns

# Rename to have consistent nomenclature of columns

df = df.rename(columns = {"Subexperiment": "Experiment", "Experiment": "Dataset"})

# Merge dataframe with morphometrics dataframe

#SOMETHING DOESN'T WORK, THERE ARE CELL CYCLES WITHOUT LINEAGE WHICH IS IMPOSSIBLE
#ALSO, reduced number of i.e. CV is because that's the one which have a lineage

sub_df = df[['Time', 'LowZoom_ID', 'Experiment_Time_mins', 'Experiment', 'Dataset',
       'Timepoint', 'Position', 'Track_ID', 'Source_ID', 'Splitting_event',
       'Lineage', 'Track_Coordinate_X', 'Track_Coordinate_Y', 'Frame',
       'Dataset', 'Generation', 'Mother_ID', 'Grandmother_ID', 'Sister_ID',
       'Seen_sister', 'Aunt_ID', 'Cousin_ID', 'Random_ID', 'Seen_granny']] ######## TEST Source_ID and Dataset and Position or so
matches_df = pd.read_csv(root + "MasterDataFrame_MatchesLineages.csv")
matches_df.columns

matches_CC_df = pd.merge(sub_df, matches_df, how = "outer")
print(matches_CC_df.shape)
matches_CC_df.head()

matches_CC_df.to_csv(root + "MasterDataFrame_MatchesLineagesWithCellCycles.csv")
print("Saved the BIG FINAL DATAFRAME.")

lineages_df.to_csv(root + "MasterDataFrame_Lineages.csv")
print("Finished compilation of concatenated lineages dataframe: {}".format(root + "MasterDataFrame_Lineages.csv"))
